KS-1-4 correct type

+++ src/main/kotlin/lesson1/lesson1_task4.kt
@@ -6,7 +6,7 @@ fun main() {
     val age: Byte = 27
     val partOfDay: Float = 0.075f
     val seconds: Short = 6480
-    val partOfYear: Float = 2.0547945205479453E-4f
+    val partOfYear: Double = 2.0547945205479453E-4
     val apogee: Int = 327000
 
     println(length)
